[IMP] conf: automatically generate the sitemap.xml for SEO
Before this commit:
Odoo's documentation does not have a sitemap.xml file. This file
it usefull to better reference the documentation accross search
engines in order for customers to find more relevant research.

After this commit:
The sitemap.xml is generated automatically for all documentation
pages automatically with link to the translated versions following
the guidelines, see:
https://developers.google.com/search/docs/specialty/international/localized-versions

The code is heavily inspired by:
https://github.com/jdillard/sphinx-sitemap
with some adaptations to fit odoo's needs